Incorporate comments 2
 - remove additional file by lumping into contract solver
- clarification of non-simple sieve title
- count primes not create full array, and don't check individual entries
